git-svn-id: file:///media/LinuxData/Development/SVN/KKEdit@1141 923dfa19-3233-4c9e-86b8-84ae44b222e5

+++ qtbranch/KKEdit/src/callbacks.cpp
@@ -366,7 +366,7 @@ int yesNo(char* question,char* file)
 VISIBLE void doOpenFile(Widget* widget,uPtr data)
 //TODO//
 {
-printf("triggered doOpenFile %i\n",(int)data);
+printf("triggered doOpenFile %i\n",(int)(long)data);
 #ifndef _USEQT5_
 	GtkWidget*	dialog;
 	char*		filename;
